Update clean_roms.py
Added support for multiple disc rom archives, if the rom is properly named related volumes will be kept during the procedure, for example if the rom has tag (Disc 1), roms with the same name but different number will be spared, same for  (Side A, Side B...) format.
Added an extra argument to costumize the list of subdirectories to ignore, for example to ignore metadata information like images, videos or manuals.
The script will now halt and give feedback to the user when unmatched parenthesis or brackets are found, for a chance to correct the culprit name.   
Fixed bugs and improved build rank functionality, '!' builds will now be prioritized over roms that dont have this sign.